+++ Source/PluginProcessor.cpp
@@ -907,8 +907,6 @@ void OdinAudioProcessor::setSampleRate(float p_samplerate) {
 	m_global_env.setSampleRate(p_samplerate);
 	m_global_lfo.setSampleRate(p_samplerate);
 	m_arpeggiator.setSampleRate(p_samplerate);
-
-	m_voice[0].start(60, 100, 60, 0, 0, 1.f, false, 0, 0);
 }
 
 void OdinAudioProcessor::initializeModules() {